restructure test to avoid linter issues
